fix: fix index just use to re-export and add TODO comment
