+++ src/pages/app/helper.ts
@@ -17,6 +17,7 @@ const useAppPrepareHook = (): IAppPrepareHook => {
       const data = await getUsers();
       setUsers(data);
     } catch (error) {
+      // TODO: this error will be handled in another branch
       console.log("🚀 ~ file: helper.ts:21 ~ getUserList ~ error:", error);
     }
   };
